RF: Dissolve custom gitrepo, annexrepo, and dataset classes
As of datalad/datalad#3183, all of the custom code from these modules
have been absorbed into the base classes.  Their remaining content is
for aliasing the core names to the Revolution* variants and for
disabling obsolete methods.  The hope is that these won't need many
changes in the future to support the revolution interfaces, so any
needed changes can be made in core.