ENH+RF: Rewrite of `aggregate-metadata()`
- This does most of what the old one did, but needs 50% less code
- Code complexity is dramatically reduced too, by not modifying a
  dataset's content until the very end and by storing metadata
  objects under a name defined by a hash of their content
  (the latter avoid a sizable number of computing cycles
  spend on rehashing in the old implementation).
- Any recording of modifications (incl. deletions) is completely
  left to rev-save()
- only datasets with an actually relevant modification are touched
  at all (avoids a lot of needless processing)

Otherwise, this is using the status()-based approach to discovering
what is there and what needs to be left alone (single call).

The code is still littered with TODOs.

Key feature that is still missing: ATM this can only do the update
strategy "target", but not "all". However, instead of being all
scrambled up into each other, "all" will be implemented as a wrapper
around this new implementation.