NF: Extractor state reporting
Extractors can now express state information, and
`extract_metadata(process_type='extractors')` can report it.

Next step is to record this state information as part of the
metadata aggregate catalog and to use it for triggering metadata
re-aggregation on state changes.

This will be instrumental in coping with future metadata layout
and vocabulary changes.

It also paves the way for extractor-selective reaggregation of metadata.